safari-600.7-branch/LayoutTests/fast/canvas/canvas-lineWidth-expected.txt

Test the default lineWidth is consistent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.lineWidth is 1
PASS ctx.lineWidth is 1
PASS ctx.lineWidth is 1
PASS ctx.lineWidth is 1
PASS ctx.lineWidth is 1
PASS ctx.lineWidth is 1
PASS successfullyParsed is true

TEST COMPLETE
